+++ CHANGELOG.md
@@ -4,12 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 2.1.0 - 2020-07-03
+### Added
+- Documentation in English and German ([#11](https://github.com/scm-manager/scm-ci-plugin/pull/11))
+- Sort Keys for Workflow Rules ([#13](https://github.com/scm-manager/scm-ci-plugin/pull/13))
+
+### Fixed
+- Workflow engine rules also consider pull request ci status for validation ([#12](https://github.com/scm-manager/scm-ci-plugin/pull/12))
+
+## 2.0.0 - 2020-06-04
+### Changed
+- Rebuild for api changes from core
+
+## 2.0.0-rc3 - 2020-05-08
 ### Added
 - Add "CI-Status All Success" rule for workflow engine ([#7](https://github.com/scm-manager/scm-ci-plugin/pull/7))
 - Implemented X CI-Status successful rule for workflow engine ([#8](https://github.com/scm-manager/scm-ci-plugin/pull/8))
 - Add swagger rest annotations to generate openAPI specs for the scm-openapi-plugin. ([#2](https://github.com/scm-manager/scm-ci-plugin/pull/5))
-- Sort Keys for Workflow Rules
+- Integrate pull request specific ci statuses ([#10](https://github.com/scm-manager/scm-ci-plugin/pull/10))
 
 ### Changed
 - Changeover to MIT license ([#6](https://github.com/scm-manager/scm-ci-plugin/pull/6))